update cypress to 9.5.3, improve ts support, remove linting, update documentation
