+++ docs/modules/quickstarters/pages/e2e-cypress.adoc
@@ -1,83 +1,104 @@
-= End-to-end tests with cypress (e2e-cypress)
+= End-to-end tests with Cypress (e2e-cypress)
 
-cypress e2e testing quickstarter project
+Cypress end-to-end testing quickstarter project
 
 == Purpose of this quickstarter
 
-This is a cypress e2e testing project quickstarter with basic setup for https://www.docker.com/[Docker], https://jenkins.io/[Jenkins], https://www.sonarqube.org/[SonarQube] and https://www.openshift.com/[OpenShift].
+This is a Cypress end-to-end testing project quickstarter with basic setup for https://www.docker.com/[Docker], https://jenkins.io/[Jenkins], https://www.sonarqube.org/[SonarQube] and https://www.openshift.com/[OpenShift].
 
 == What files / architecture is generated?
 
 ----
 .
+├── fixtures
+│   └── example.json
+│── plugins
+│   └── index.js
+│── support
+│   ├── commands.ts
+│   ├── generic-login.ts
+│   ├── index.ts
+│   ├── msalv2-login.ts
+│   └── test-evidence.ts
+├── tests
+|   ├── acceptance
+│   |   └── acceptance.spec.ts
+|   ├── installation
+│   |   └── installation.spec.ts
+|   └── integration
+│       └── integration.spec.ts
+├── cypress-acceptance.json
+├── cypress-installation.json
+├── cypress-integration.json
+├── cypress.env.json.template
+├── cypress.json
 ├── Jenkinsfile
-├── README.md
-├── files
-│   ├── README.md
-│   ├── build
-│   |   ├── tests
-|   |   |   ├── acceptance
-|   │   │   |   └── acceptance.spec.js
-|   |   |   ├── installation
-|   │   │   |   └── installation.spec.js
-|   |   |   └── integration
-|   │   │       └── integration.spec.js
-│   |   ├── test-results
-|   |   |   ├── acceptance-junit.xml
-|   |   |   ├── installation-junit.xml
-|   |   |   └── integration-junit.xml
-│   │   └── support
-│   │       ├── commands.js
-│   │       └── index.js
-│   ├── cypress-acceptance.json
-│   ├── cypress-installation.json
-│   ├── cypress-integration.json
-│   ├── fixtures
-│   │   └── example.json
-│   ├── tests
-|   |   ├── acceptance
-│   │   |   └── acceptance.spec.ts
-|   |   ├── installation
-│   │   |   └── installation.spec.ts
-|   |   └── integration
-│   │       └── integration.spec.ts
-│   ├── package.json
-│   ├── plugins
-│   │   └── index.js
-│   ├── support
-│   │   ├── commands.ts
-│   │   └── index.ts
-│   ├── tsconfig.json
-│   └── tslint.json
-├── init.sh
 ├── metadata.yml - Component metadata
-└── release-manager.yml - Configuration file for the Release Manager
+│── package.json
+├── README.md
+│── release-manager.yml - Configuration file for the Release Manager
+└── tsconfig.json
 ----
 
 == Frameworks used
 
-This project is generated by https://cli.angular.io/[Angular CLI]
+* https://www.cypress.io[Cypress]
+* https://www.typescriptlang.org[TypeScript]
 
-******* http://www.cypress.io[cypress]
+== Usage - how do you start after you provisioned this quickstarter
 
-******* http://www.typescriptlang.org/[TypeScript]
-
-## Usage - how do you start after you provisioned this quickstarter
-
-Check the README.md file at root level for further instructions after the quickstarter has been provisioned.
+Check the `README.md` file at root level for further instructions after the quickstarter has been provisioned.
 
 == How this quickstarter is built through jenkins
 
-The Jenkinsfile is provisioned with this quickstarter to ease CI/CD process.
-In Jenkinsfile, there is the following stage:
+The `Jenkinsfile` is provisioned with this quickstarter to ease CI/CD process. In `Jenkinsfile` there is the following stage:
 
-* stageTest - Lint the TypeScript code by running `npm install` and `npm run lint`, and transpiles it to JavaScript and run the tests with the `npm run build` command.
+stageTest:: Load Node.js dependencies by running `npm install` and executing the e2e tests by running `npm run e2e`.
+
+Please note: *TSLint* has been deprecated in favor of *ESLint*. Therefore the support for TSLint has been removed from this quickstarter. Please consider adding ESLint support (see also https://www.npmjs.com/package/eslint#installation-and-usage) or a formatter like *Prettier* (see also https://prettier.io/docs/en/install.html).
+
+=== Cypress without post-installation
+
+While `npm install` is gathering the dependencies it is also performing a post-installation process which is downloading Cypress' binaries. It is possible to change this behavior, e.g. for getting those binaries from Nexus or a local cache. See https://docs.cypress.io/guides/getting-started/installing-cypress#Advanced for more details.
+
+A more advanced solution could be the use of a custom builder agent, where Cypress is already embedded. The related `Dockerfile` could look like this:
+
+[source,docker]
+----
+# The following FROM tag is informational. It is overwritten by OpenShift.
+FROM ods/jenkins-agent-nodejs16:4.x
+
+ARG cypressVersion=9.5.3
+ARG cypressHash=...
+
+USER root
+
+# Adds cypress@VERSION binaries to /home/jenkins/.cache/Cypress/VERSION/Cypress
+# This allows to disable post-installation of Cypress binaries and gather those binaries from a cache folder instead. The cache
+# folder holds the binaries for the respective versions of Cypress.
+# Setting the following environment variable in the Jenkinsfile allows to disable post-installation of cypress@...:
+#   CYPRESS_INSTALL_BINARY=0
+# The custom cache folder can be configured in Jenkinsfile by the following environment variable setting:
+#   CYPRESS_CACHE_FOLDER=/home/jenkins/.cache/Cypress
+ADD https://download.cypress.io/desktop/${cypressVersion}?platform=linux&arch=x64 /home/jenkins/cypress/cypress-linux.zip
+RUN md5sum /home/jenkins/cypress/cypress-linux.zip && \
+    echo "${cypressHash} /home/jenkins/cypress/cypress-linux.zip" | md5sum -c && \
+    mkdir -p /home/jenkins/.cache/Cypress/${cypressVersion} && \
+    unzip -q /home/jenkins/cypress/cypress-linux.zip -d /home/jenkins/.cache/Cypress/${cypressVersion} && \
+    rm /home/jenkins/cypress/cypress-linux.zip
+
+# fix access rights
+RUN chgrp -R 0 $HOME && \
+    chmod -R g=u $HOME && \
+    chmod ug=rx /home/jenkins/.cache/Cypress/${cypressVersion}/Cypress/Cypress
+USER 1001
+----
 
 == Builder agent used
 
 This quickstarter uses
-https://github.com/opendevstack/ods-quickstarters/tree/master/common/jenkins-agents/nodejs12[Nodejs12 builder agent] Jenkins builder agent.
+https://github.com/opendevstack/ods-quickstarters/tree/master/common/jenkins-agents/nodejs16[Node.js 16 builder agent] for Jenkins.
 
 == Known limitations
 
-For now the URL of the front end instance to test is hard-coded inside the `package.json` file. There should be a better solution for handing this over. It is also possible to run `npm run e2e-at` with an additional parameter, e.g. `+npm run e2e-at https://domain.com+`.
+NA
