+++ e2e-cypress/files/.prettierrc
@@ -1,0 +1,8 @@
+{
+  "arrowParens": "avoid",
+  "bracketSpacing": true,
+  "endOfLine": "auto",
+  "printWidth": 140,
+  "singleQuote": true,
+  "trailingComma": "es5"
+}
