+++ e2e-cypress/files/cypress-installation.json
@@ -1,14 +1,14 @@
 {
-  "baseUrl": "http://www.example.com",
+  "baseUrl": "https://www.w3schools.com",
   "fixturesFolder": "fixtures",
-  "integrationFolder": "build/tests/installation",
+  "integrationFolder": "tests/installation",
   "pluginsFile": "plugins/index.js",
-  "supportFile": "build/support",
+  "supportFile": "support",
   "viewportWidth": 1376,
   "viewportHeight": 660,
   "video": false,
   "reporter": "junit",
   "reporterOptions": {
-    "mochaFile": "build/test-results/installation-junit.xml"
+    "mochaFile": "test-results/installation-junit.xml"
   }
 }
