+++ e2e-cypress/files/cypress.env.json.template
@@ -1,0 +1,7 @@
+{
+  "TENANT_ID": "",
+  "CLIENT_ID": "",
+  "CLIENT_SECRET": "",
+  "USERNAME": "",
+  "PASSWORD": ""
+}
