+++ e2e-cypress/files/metadata.yml
@@ -1,6 +1,6 @@
 ---
 name: Cypress
-description: "Cypress is a JavaScript-based end-to-end testing framework that doesn't use Selenium at all. It is built on top of Mocha, which is again a feature-rich JavaScript test framework running on and in the browser, making asynchronous testing simple and fun. Technologies: Cypress 3.6, TypeScript 3.1"
+description: "Cypress is a JavaScript-based end-to-end testing framework. It is built on top of Mocha, which is again a feature-rich JavaScript test framework running on and in the browser, making asynchronous testing simple and fun. Technologies: Cypress 9.5, TypeScript 4.6"
 supplier: https://www.cypress.io
 version: 4.x
 type: ods-test
