+++ e2e-cypress/files/package.json
@@ -1,30 +1,23 @@
 {
-  "name": "e2e",
+  "name": "e2e-cypress",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
-    "lint": "tslint --project .",
-    "build": "tsc --project .",
-    "watch": "npm-run-all --parallel \"build -- --watch\" \"cypress:open\"",
+    "start": "npm run cypress:open",
     "cypress:open": "cypress open",
-    "cypress:run-installation": "cypress run --config-file cypress-installation.json --browser chrome",
-    "cypress:run-integration": "cypress run --config-file cypress-integration.json --browser chrome",
-    "cypress:run-acceptance": "cypress run --config-file cypress-acceptance.json --browser chrome",
+    "cypress:run": "cypress run --browser=chrome",
+    "cypress:run-installation": "npm run cypress:run -- --config-file cypress-installation.json",
+    "cypress:run-integration": "npm run cypress:run -- --config-file cypress-integration.json",
+    "cypress:run-acceptance": "npm run cypress:run -- --config-file cypress-acceptance.json",
     "delete-junit-results": "rimraf build/test-results",
-    "e2e": "npm-run-all lint build delete-junit-results cypress:run-installation cypress:run-integration cypress:run-acceptance"
+    "e2e": "npm-run-all delete-junit-results cypress:run-installation cypress:run-integration cypress:run-acceptance"
   },
   "private": true,
-  "dependencies": {
-    "rxjs": "^7.1.0"
-  },
   "devDependencies": {
-    "@types/node": "^15.12.2",
-    "cypress": "^7.5.0",
+    "@types/node": "^17.0.23",
+    "cypress": "^9.5.3",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2",
-    "rxjs-tslint": "^0.1.8",
-    "ts-node": "^10.0.0",
-    "tslint": "^5.11.0",
-    "typescript": "^4.3.2"
+    "typescript": "^4.6.3"
   }
 }
