+++ e2e-cypress/files/support/index.ts
@@ -13,17 +13,4 @@
 // https://on.cypress.io/configuration
 // ***********************************************************
 
-// Import commands.js using ES2015 syntax:
 import './commands';
-
-declare global {
-    namespace Cypress {
-        interface Chainable {
-            msalv2Login(): void;
-            crowdLogin(asTest: boolean): void;
-        }
-    }
-}
-
-// Alternatively you can use CommonJS syntax:
-// require('./commands')
