+++ e2e-cypress/files/tests/installation/installation.spec.ts
@@ -1,10 +1,9 @@
-describe('installation e2e tests', function () {
-
-  it('Application is reachable and does not show error page', function () {
+describe('installation e2e tests', () => {
+  it('Application is reachable and does not show error page', () => {
+    cy.login();
     // cy.msalv2Login();
     // cy.visit('');
     // cy.get('app-header').should('be.visible');
     // cy.get('app-error-page').should('not.exist');
   });
-
 });
