+++ e2e-cypress/files/tests/integration/integration.spec.ts
@@ -1,4 +1,4 @@
 /* tslint:disable:no-unused-expression */
-describe('integration e2e tests', function () {
-    // see installation.spec.ts for examples
+describe('integration e2e tests', () => {
+  // see installation.spec.ts for examples
 });
