+++ e2e-cypress/files/tsconfig.json
@@ -1,24 +1,8 @@
 {
-  "include": [
-    "tests/**/*.spec.ts",
-    "support/*.ts",
-    "plugins/*.ts"
-  ],
   "compilerOptions": {
-    "baseUrl": "./",
-    "outDir": "build",
-    "sourceMap": false,
-    "declaration": false,
-    "moduleResolution": "node",
     "target": "es5",
-    "skipLibCheck": true,
-    "typeRoots": [
-      "node_modules/@types"
-    ],
-    "lib": [
-      "es2017",
-      "dom"
-    ],
+    "lib": ["es2017", "dom"],
     "types": ["cypress", "node"]
-  }
+  },
+  "include": ["**/*.ts"]
 }
